added home screen and set buttons
