+++ src/App.js
@@ -1,10 +1,11 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "./logo.svg";
+import "./App.css";
+import HomeScreen from "./Screens/HomeScreen/HomeScreen";
 
 function App() {
   return (
     <div className="App">
-      I am text
+      <HomeScreen />
     </div>
   );
 }
