+++ src/Screens/EditSetScreen/EditSetScreen.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const EditSetScreen = () => {
+  return <div>Hi from edit screen</div>;
+};
+
+export default EditSetScreen;
